lrdp: iommulib: Add support for using only iommu_debug_attachments
iommu_debug_attachments now has all of the information that is necessary
for generating the S1 page tables for an IOMMU domain. So, split the
existing code into two parts: the legacy part, which will collect
the information needed to generate the S1 page tables for an IOMMU
domain, and the new part which uses only the information provided in
the iommu_debug_attachments structure.

Change-Id: I4c94e74a556d3c5ef69207d535efe609d80b71f9