lrdpv2: ftrace_event: Fix dynamic string extraction
For events with entry of "__data_loc char[]" type, the value
is actually composed of (len << 16 | offset). Use this second
offset (typically located after the rest of the entries) to properly
read the string from the event data.

Also change the fmt_name_value_map to an OrderedDict so that
the event entries are printed out in the original order.

Finally optimize the looping over offset_data entries by using the
.items() iterator to get both key and value rather than enumerating
with .keys() and a separate lookup.

Change-Id: Id7ed13daf53e9ab2734062ba759a0746b0f521e3