lrdp_v2 : Suppress windows prompt during gdb crash
Change suppresses the windows prompt generated by a crash
in gdb exe which results in ramparser getting stuck
waiting for a button press till it times out. On
timeout all the successive subparsers also fail.
This change catches the exception generated by gdb
crash but doesn't allow the prompt to be displayed.

Change-Id: Ie58111fc586fa3587ad1cadd0645c0173cc3c86b