lrdp_v2: Add error handling while extracting sched info for active cpus
Ramparser was failing if we are unable to extract sched info for active
cpu's due to any reason. sub-parsers of ramparser were not getting executed
as they are called after checking sched info for active cpu's. So added
error handling to handle this and not stop ramparser in case of failure.

Change-Id: I3c940b96f266c61e441763c0b20e9f6e6db4b039