lrdpv2: Decode function pointers to a CFI jump table
When CFI is enabled for a kernel, function pointers point to a jump
table which contains the real destination. The closest symbol in
the jump table may not be meaningful (e.g.
__typeid__ZTSFyvE_global_addr) and parsers like timerlist check
the function pointer symbol to unpack further (i.e. deferred_work_fn).

Create a new function, function_pointer_lookup, which walks through a
CFI jump table when a function pointer points to a CFI jump table entry.
To figure out the destination of the jump table, disassemble the entry
and figure out the branch destination address.

Change-Id: I1d73166113e2440de2452141a6652ac492aefa0c