ramdump.py: Parse KMODULES file for minidump
In minidump, modules linked list isn't available and instead a KMODULES
file is created which lists the module name and base address in textual
format:

  name: qcom_scm, base: 0xfffffff007910000 size: 0x1c000

Parse this file to determine loaded modules and their base address and
create a basic module table. Minidump is missing kallsyms info, but this
information isn't required for basic parsing.

Change-Id: Ib7f474c36e4ea52df4b5352c3be6f258bb5be115