parsers: rtb: Add support for new RTB layout in minidump
New RTB layout moves all RTB data to one contiguous location, but the
pointer to the region isn't registered for minidump when RTB is a module
since .data from modules aren't collected by default. This entire region
is registered under KRTB_BUF name, so search for a section named
KRTB_BUF when parsing a minidump and use that for the RTB buffer
directly.

Change-Id: I0a10823e46e00d06e37b3fef7e8e0bb41a787a0c