ramdump: Add support to read memory from elffile
In minidump, not all memory sections are dumped. Some of these memory
sections are const data from vmlinux or KOs, and could be loaded
directly from ELF since symbols match.

GDB supports dumping memory from ELF with "dump binary memory" command
into a temporary file. Add support in the RamDump.read_* functions to
allow the data to come from elf file, and default this support to
"False" to avoid breaking things. In case the memory is not part of
".data" section, then it would be part of BSS and return zeroes instead
of None.

Change-Id: Ibd41be8b05fb3ac1878131608ff1b430da5ca112