lrdp: ramdump: Parse only properly live modules
Modules which aren't completely live yet may be in temporary state that
is not consistent and cause LRDP to hang since kallsyms not yet
developed and module section list not yet complete. Check that module
state is LIVE and additional check before loading module sections that
the sysfs node for sections is really "sections." These checks were
found to be sufficient when loading a crash where modules are still in
progress of being loaded.

Change-Id: I339ce3b15455eac9321b43b0574cd2f0e8404e7b