ramdump: Initial support >v5.15
Add initial support for v5.15 and newer kernels. Basic modules parsing
and dmesg was validated. Only one change impacted ramdump parser:
https://lore.kernel.org/all/20201008153602.9467-3-ardb@kernel.org/.

The KASAN shadow region is reconfigured so that it ends at the start of
the vmalloc region, and grows downwards. That way, the arrangement of
the vmalloc space (which contains kernel mappings, modules, BPF region,
the vmemmap array etc) is identical between non-KASAN and KASAN builds,
which aids debugging.

This impacted our offline calculation of kimage_vaddr which is used to
determine virtual and physical offsets.

Change-Id: I0e8b2d98db9d817f301fc2819948118ab5acddf5