Fix SECTION_SIZE_BITS for non-android kernels
commit f0b13ee23241 ("arm64/sparsemem: reduce SECTION_SIZE_BITS")
was merged into upstream v5.12 kernel and backported to android
kernels. But upstream stable kernels do not have this backport since
it is not a fix but a perf improvement of some sort. So for other
kernels such as chrome which is based on upstream kernel, lrdp
section size bits is not proper thereby breaking pagetracking and
needs to be set to older value of 30. So fix this by adding another
starting check for "MEMORY_HOTPLUG=n" to set the section size bits
to 30 since memory hotplug was the primary usecase for setting a
lower section size bits.

Signed-off-by: Bhuvaneshwari Balagar <quic_c_bhuvba@quicinc.com>

Change-Id: I88c4522bf93b39ab29b6a51d3739f5c6522c853b